requirements.txt
Django==5.0.3
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0